fix(sql): a suffix-less floating point literal is a Double, not a Float (#7609)
A floating point literal written without a type suffix was parsed into a Float,
and the comparison operators then widened that Float back with .doubleValue(),
which reproduces the single precision rounding error as a double instead of
removing it. Against a DOUBLE or DECIMAL property the literal was therefore not
the number the user typed, and every record sitting exactly on the boundary was
decided the wrong way, with no error and no warning:

  disc >= 0.05   returned what `> 0.05` returns
  disc =  0.05   returned nothing
  disc <  0.05   returned the records that are EQUAL to 0.05

Which side loses depends only on the sign of the rounding error of that literal,
so it is not always the lower bound: at 0.7 the float form rounds down and it is
the upper bound that drops its records. BETWEEN, a bound parameter, an explicit
D suffix and an indexed column were all correct on the same data, so adding or
dropping an index changed the answer of identical SQL. Found on a TPC-H Q6
shape, where it summed two of the three discount buckets and reported revenue
about 30% low.

Two halves, both fixed:

- The literal. FloatingPoint.getValue() and the ANTLR builder's float literal
  visitor now read a suffix-less literal as a Double, matching Java, Cypher and
  ArcadeDB's own D suffix. The F suffix still asks for single precision. This
  also removes a second defect in the ANTLR path, where a magnitude above
  Float.MAX_VALUE silently became infinity.

- The comparator. Type.castComparableNumber() widened a Float to Double with
  .doubleValue() and built a BigDecimal with BigDecimal.valueOf(float), both of
  which carry the float's rounding error into the wider type. They now go
  through the decimal form, which is what Type.convert() already does for the
  index path - the two have to agree or the same query answers differently
  depending on whether an index exists. The mapping is strictly monotonic, so
  the comparison stays a total order, and the round-trip is skipped for the
  non-finite values and for an integral float below 2^24, which widens exactly.

Two spellings that were already correct keep their type across the change:
InputParameter renders a bound Float with its F suffix, and PNumber.toString()
emits the suffix too, so rendering a statement and parsing it back (EXPLAIN, the
statement cache key, a rewritten sub-query) no longer changes what it computes.

Four existing tests asserted the old Float typing of a suffix-less literal and
now assert the new one, with the F-suffixed spelling kept alongside where the
point of the test was the FLOAT type itself.

+++ engine/src/test/java/com/arcadedb/function/sql/sql/SQLFunctionsTest.java
@@ -264,7 +264,8 @@ class SQLFunctionsTest {
 
     assertThat(myresult.remove(Integer.valueOf(1))).isTrue();
     assertThat(myresult.remove(Integer.valueOf(4))).isTrue();
-    assertThat(myresult.remove(Float.valueOf(5))).isTrue();
+    // ISSUE #7609: A SUFFIX-LESS LITERAL IS A DOUBLE, SO `5.00` IS A Double
+    assertThat(myresult.remove(Double.valueOf(5))).isTrue();
     assertThat(myresult.remove("john")).isTrue();
 
     assertThat(myresult.size()).isEqualTo(1);
